Reconfigured test-cmd to continue on failure
As test-cmd test bucket sizes shrink and cross-test
dependencies are removed, it makes less and less
sense to fail once and for all when a test-cmd test
case fails. Instead, we know we can't run any more
tests in the same bucket as the failed test, but we
can try to run every other test bucket.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>